Redesign Pre-Install Main View
